Replace C++ example with a somewhat realistic use case for mocks
The example directory contains a simple library for working
with "projects" which may contain a `settings.toml` file, an executable
consuming this library, and a GTest-based test suite which unit tests
the behavior of the the `Project` class.

The backing store for the projects is dependency-injected into the
`Project` class and must implement the `ProjectStorage` interface. The
executable uses the "real" implementation which read from and writes to
the filesystem, while the test suite uses a mock onto which it sets
expectations to verify the correct behavior of the `Project` class
without side effects. The mock class definition `ProjectStorageMock.h`
can be created using mockery from the interface:

    mockery create example/src/project/Project.cpp -i ProjectStorage > example/test/ProjectStorageMock.h