Introduce application struct `MockeryApp` & custom result/error types
The main function has also been split into several functions:
- `Mockery::new` takes care of parsing the translation unit.
- `cli_main` takes on the duties which would usually be carried out by
  a `main` function, i.e. it handles CLI parsing, constructs the app,
  and then runs the appropriate subcommand.
- `main` just calls main and handles error reporting.

The reason for the split between `main` and `cli_main`, is that even
though `main` may return a `Result<>`, Rust does not offer any way to
customize the error reporting, but prefixes the `Debug` string
with "Error: ". It will be possible to customize this, once the
`std::process::Termination` trait is stabilized. When it lands in stable
Rust, we can do away with the current `main` and simply renaming
`cli_main` to `main`. The custom error reporting can then be done in
the implementation of the `Termination` trait for `CLIResult`,
specifically in the `report()` function. Until then, the split is
necessary to ensure that all the stack variables in `cli_main` is
guaranteed to be cleaned up by the time that `std::process::exit` is
called, as this does not do any further cleanup.