Filter out `/TP` et al. from the compile command args
These options are used by MSVC to set the language to C (`/Tc`, `/TC`)
or C++ (`/Tp`, `/TP`). When a compile_commands.json database which
targets the MSVC toolchain is fed into libclang, it supplements the
arguments by `--driver-mode=cl` automagically. However, as soon as one
of `/TP` or its friends is present in the subsequent arguments, the
parser comes back with an `AstDeserialization` error. Luckily, these
options are not necessary at all, as `--driver-mode=cl` seems to
implicitly turn on the C++ mode as it stands. Therefore, the offending
set of args is just filtered out.