Turn `CLIError` into an enum with one variant per error type
The formatting of the actual error message now happens only in the
`Display` implementation for `CLIError`. This is both more efficient
(eliding heap allocations for temporary `String` fragments but writing
straight to the formatter), and less distracting at the call site while
still conveying the benefits of addition context from the caller.

It also allows dedicated handling of certain error types using match
expressions. This is akin to being able to catch dedicated custom
exceptions in languages with exception-based error handling like C++.