Filesystem iterator now begins search for compile commands database at the source dir
Previously, it went up `<search-radius>` many directories and then
performed a standard breath-first search. This could result in the
wrong compile commands database (e.g. one belonging to a different
project in an adjacent directory) being picked up first. The new
behavior reduces the chance of something like that happening by
searching the filesystem hierarchy starting from the source file and
moving on from there successively. This is similar to how degrees of
kinship work: First-degree relatives (parents and children) are
considered first, followed by second-degree relatives (cousins and
grandparents), and so on.

The default value of the search radius has been upped to 3 because the
counting is more rigorous now and it is more likely that the search
terminates quite early should the database be closer than that.
Otherwise the default value would not have worked for the example
project.